Styling and naming changes * Changed the name of the demo to the correct component * Removed InlineStyleSheet to instead use the Size API
